Separate test cases and correct github actions.
