+++ .github/workflows/ci_env.yml
@@ -6,28 +6,28 @@ jobs:
   tests:
     runs-on: ubuntu-22.04
     strategy:
-      max-parallel: 5
+      matrix:
+        python-version: ["3.8"]
 
     steps:
     - name: Checkout github repo
-      uses: actions/checkout@v3
-    - uses: conda-incubator/setup-miniconda@v2
+      uses: actions/checkout@v4
+    - name: Setup Miniconda
+      uses: conda-incubator/setup-miniconda@v3
       with:
         auto-update-conda: true
         python-version: 3.8
+        miniforge-version: latest
+
     - name: Conda info
       shell: bash -el {0}
       run: conda info
-    - name: Install python and gcc
+
+    - name: Install system dependencies
       shell: bash -el {0}
       run: |
-        conda install -c conda-forge python=3.8
-        conda install -c conda-forge gcc=12.1.0
-    - name: Install gdal eccodes hdf5
-      shell: bash -el {0}
-      run: |
-        conda install -c conda-forge "gdal<=3.9.2" eccodes
-        conda install -c conda-forge hdf5
+        conda install -c conda-forge "gcc=12.1.0" "gdal<=3.9.2" eccodes hdf5
+
     - name: Install dependencies
       shell: bash -el {0}
       run: |
@@ -39,6 +39,7 @@ jobs:
         gdal-config --version
         python -c "from osgeo import gdal; print(gdal.__version__)"
         conda list
+
     - name: Test with pytest
       shell: bash -el {0}
       run: |
