+++ pytest.ini
@@ -1,5 +1,5 @@
 [pytest]
-python_files = *_tests.py
+python_files = tests_*.py
 python_classes = Test
 python_functions = test_*
 norecursedirs = build
