+++ tests/tests_efas.py
@@ -136,44 +136,3 @@ class TestEFAS1arcmin360DaysCalendar(TestLis):
 
     def test_et(self):
         return self.listest('et')
-
-
-class TestCORDEX(TestLis):
-    domain = 'cordex'
-    settings_path = os.path.join(current_dir, 'data/tests_cordex.xml')
-
-    def test_e0(self):
-        return self.listest('e0')
-
-    def test_et(self):
-        return self.listest('et')
-
-
-class TestGLOFAS(TestLis):
-    domain = 'glofas'
-    settings_path = os.path.join(current_dir, 'data/tests_glofas.xml')
-
-    def test_e0(self):
-        return self.listest('e0')
-
-
-class TestRel_Humidity_360_Cal(TestLis):
-    domain = 'rel_humidity_360_cal'
-    settings_path = os.path.join(current_dir, 'data/tests_rel_humidity_360_cal.xml')
-
-    def test_et(self):
-        return self.listest('et')
-
-    def test_tair(self):
-        return self.listest('tair')
-
-
-class TestHargreaves(TestLis):
-    domain = 'hargreaves'
-    settings_path = os.path.join(current_dir, 'data/tests_hargreaves.xml')
-
-    def test_et(self):
-        return self.listest('et')
-
-    def test_tair(self):
-        return self.listest('tair')
